requirements.txt
torch
torchaudio
numpy
pandas
matplotlib
scikit-learn
faiss-cpu
transformers
librosa